updates on protocol conf, pdb dim red, protocol genesis
