+++ continuousflex/protocols.conf
@@ -101,6 +101,9 @@ Genesis = [
 	{"tag": "section", "text": "3. Energy Minimization", "children": [
 	    {"tag": "protocol", "value": "ProtGenesis", "text": "GENESIS", "icon": "bookmark.png"}
 	]},
+	{"tag": "section", "text": "4. Normal Mode Analysis (Optional)", "children": [
+        {"tag": "protocol", "value": "FlexProtNMA", "text": "NMA"}
+	]},
 	{"tag": "section", "text": "4. Flexible Fitting using MD / NMMD", "children": [
 	    {"tag": "protocol", "value": "ProtGenesis", "text": "GENESIS", "icon": "bookmark.png"}
 	]}]
